move: src path로 이동
